+++ Backend/src/main/java/com/codementor/repository/ArticleRepository.java
@@ -11,21 +11,21 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<Article, Long> {
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
     Optional<Article> findBySlug(String slug);
     
     @Query("SELECT a FROM Article a WHERE a.isApproved = true")
     Page<Article> findAllApproved(Pageable pageable);
     
     @Query("SELECT a FROM Article a WHERE a.track.id = ?1 AND a.isApproved = true")
-    Page<Article> findByTrackId(Long trackId, Pageable pageable);
+    Page<Article> findByTrackId(Integer trackId, Pageable pageable);
     
     @Query("SELECT a FROM Article a WHERE a.topic.id = ?1 AND a.isApproved = true")
-    Page<Article> findByTopicId(Long topicId, Pageable pageable);
+    Page<Article> findByTopicId(Integer topicId, Pageable pageable);
     
     @Query("SELECT a FROM Article a WHERE a.subtopic.id = ?1 AND a.isApproved = true")
-    Page<Article> findBySubtopicId(Long subtopicId, Pageable pageable);
+    Page<Article> findBySubtopicId(Integer subtopicId, Pageable pageable);
     
     @Query("SELECT a FROM Article a JOIN a.jobRoles j WHERE j.id = ?1 AND a.isApproved = true")
-    Page<Article> findByJobRoleId(Long jobRoleId, Pageable pageable);
+    Page<Article> findByJobRoleId(Integer jobRoleId, Pageable pageable);
 }
