+++ Backend/src/main/java/com/codementor/repository/JobRoleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobRoleRepository extends JpaRepository<JobRole, Long> {
+public interface JobRoleRepository extends JpaRepository<JobRole, Integer> {
 }
