+++ Backend/src/main/java/com/codementor/repository/TrackRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TrackRepository extends JpaRepository<Track, Long> {
+public interface TrackRepository extends JpaRepository<Track, Integer> {
     Optional<Track> findByName(String name);
     boolean existsByName(String name);
 }
